fix testutils dependency and move model+graphql integration
