+++ aws-api/src/main/java/com/amplifyframework/api/graphql/model/ModelPagination.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package com.amplifyframework.core.model.graphql;
+package com.amplifyframework.api.graphql.model;
 
 import androidx.annotation.Nullable;
