+++ testutils/src/main/java/com/amplifyframework/testutils/sync/SynchronousApi.java
@@ -26,9 +26,9 @@ import com.amplifyframework.api.rest.RestOptions;
 import com.amplifyframework.api.rest.RestResponse;
 import com.amplifyframework.core.async.Cancelable;
 import com.amplifyframework.core.model.Model;
-import com.amplifyframework.core.model.graphql.ModelMutation;
-import com.amplifyframework.core.model.graphql.ModelQuery;
-import com.amplifyframework.core.model.graphql.ModelSubscription;
+import com.amplifyframework.api.graphql.model.ModelMutation;
+import com.amplifyframework.api.graphql.model.ModelQuery;
+import com.amplifyframework.api.graphql.model.ModelSubscription;
 import com.amplifyframework.core.model.query.predicate.QueryPredicate;
 import com.amplifyframework.testutils.AmplifyDisposables;
 import com.amplifyframework.testutils.Await;
